module04 ex03 : fixed segfault

+++ module02/ex02/Fixed.cpp
@@ -111,7 +111,7 @@ Fixed Fixed::operator-(const Fixed obj) const {
 Fixed Fixed::operator*(const Fixed obj) const {
     Fixed newObj;
 
-    newObj._fixed = (this->_fixed * obj._fixed) >> this->_fraction;
+    newObj._fixed = ((long)this->_fixed * (long)obj._fixed) >> this->_fraction;
     return (newObj);
 }
